- Now using ng-module-sort to organise components. - components organised by module. - Library files now merged into a single large file.
