+++ .gitignore
@@ -1,3 +1,4 @@
+/.tmp
 bower_components
 .idea
 node_modules
