+++ src/angularModule.js
@@ -1,16 +1,16 @@
-(function(){
+'use strict';
 
-  'use strict';
+window.cancelAnimationFrame = window.cancelAnimationFrame ||
+  window.webkitCancelAnimationFrame ||
+  window.mozCancelAnimationFrame;
 
-  window.cancelAnimationFrame = window.cancelAnimationFrame ||
-    window.webkitCancelAnimationFrame ||
-    window.mozCancelAnimationFrame;
+window.requestAnimationFrame = window.requestAnimationFrame ||
+  window.webkitRequestAnimationFrame ||
+  window.mozRequestAnimationFrame;
 
-  window.requestAnimationFrame = window.requestAnimationFrame ||
-    window.webkitRequestAnimationFrame ||
-    window.mozRequestAnimationFrame;
-
-  angular.module('angularAudioRecorder', [
-
-  ]);
-})();
+angular.module('angularAudioRecorder', [
+  'angularAudioRecorder.config',
+  'angularAudioRecorder.services',
+  'angularAudioRecorder.controllers',
+  'angularAudioRecorder.directives'
+]);
