+++ src/controllers/module.js
@@ -1,0 +1,6 @@
+'use strict';
+
+angular.module('angularAudioRecorder.controllers', [
+  'angularAudioRecorder.config',
+  'angularAudioRecorder.services'
+]);
