+++ src/directives/module.js
@@ -1,0 +1,7 @@
+'use strict';
+
+angular.module('angularAudioRecorder.directives', [
+  'angularAudioRecorder.config',
+  'angularAudioRecorder.services',
+  'angularAudioRecorder.controllers'
+]);
