+++ src/services/module.js
@@ -1,0 +1,1 @@
+angular.module('angularAudioRecorder.services', ['angularAudioRecorder.config']);
